remove dependency lines from CI config
The gensim installer should be able to handle this stuff by itself. I'm
pretty sure these version pins are are a work-around for an issue that
doesn't affect us anymore. More specifically, we used to build wheels
using older Python versions that did not have the corresponding numpy
wheels. We dropped support for older Python versions, so that should not
affect us anymore.